♻️ move site emails props into a presenter
