+++ app/views/sites/emails/edit.html.slim
@@ -56,4 +56,4 @@ div class="pf-l-flex pf-m-column"
                 label class="pf-c-form__label"
                   span class="pf-c-form__label-text"
                     = t('.exceptions_label')
-              div id="custom-support-emails" class="pf-c-form__group-control pf-m-stack" data-props=props.to_json
+              div id="custom-support-emails" class="pf-c-form__group-control pf-m-stack" data-props=presenter.props.to_json
